+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/signaling/message/Catalog.kt
@@ -35,7 +35,8 @@ data class VideoSetting(
 )
 
 data class AudioSetting(
-        @SerializedName("codec_type") val codecType: String?
+        @SerializedName("codec_type") val codecType: String?,
+        @SerializedName("bit_rate")   var bitRate:   Int?    = null
 )
 
 data class IceServer(